size調整

+++ Core/Core.Ore/Config/OreConfigData.cs
@@ -16,17 +16,17 @@ namespace Core.Ore.Config
         [DataMember(Name = "oreId")] private int _oreId;
         [DataMember(Name = "name")] private string _name;
         [DataMember(Name = "miningItemId")] private int _miningItem;
-        [DataMember(Name = "veinSize")] private int _veinSize;
-        [DataMember(Name = "veinFrequency")] private int _veinFrequency;
+        [DataMember(Name = "veinSize")] private float _veinSize;
+        [DataMember(Name = "veinFrequency")] private float _veinFrequency;
         [DataMember(Name = "priority")] private int _priority;
 
         public int Priority => _priority;
 
         public int MiningItem => _miningItem;
 
-        public int VeinSize => _veinSize;
+        public float VeinSize => _veinSize;
 
-        public int VeinFrequency => _veinFrequency;
+        public float VeinFrequency => _veinFrequency;
 
         public int OreId => _oreId;
